+++ packages/frontend/src/config/index.ts
@@ -1,3 +1,5 @@
 export const Port = 3000;
 
 export const CakeApiUrl = 'http://localhost:3100/cakes';
+
+export const CakeFormName = 'CakeForm';